fix: address code review - remove i18n alignment spaces, clean up cast comparison
Agent-Logs-Url: https://github.com/sublimit1736/term4k_sub/sessions/06ee90b9-426a-40c4-884d-2fa6641361e4

+++ src/main/ui/GameplayUI.cpp
@@ -636,7 +636,7 @@ ftxui::Component GameplayUI::component(ftxui::ScreenInteractive &screen,
                     for (const auto &[kc, ts] : state->heldKeys) {
                         const auto elapsed =
                             std::chrono::duration_cast<std::chrono::milliseconds>(now - ts).count();
-                        if (elapsed > static_cast<long long>(kKeyHoldTimeoutMs)) {
+                        if (elapsed > kKeyHoldTimeoutMs) {
                             toRelease.push_back(kc);
                         }
                     }
